valueaccounting.faircoin_utils.py is merged into faircoin.utils.py
