+++ faircoin/process_faircoin_requests.py
@@ -15,24 +15,6 @@ from valuenetwork.valueaccounting.lockfile import FileLock, AlreadyLocked, LockT
 
 #FAIRCOIN_DIVISOR = int(1000000)
 
-def init_electrum_fair():
-    try:
-        daemon = efn.daemon_is_up()
-    except:
-        logger.critical("Cannot connect with daemon.")
-        return False
-
-    if not daemon or (daemon == 'ERROR'):
-        return False
-
-    try:
-        network = efn.is_connected()
-    except:
-        logger.critical("Cannot connect with electrum-server.")
-        return False
-
-    return network and (network != 'ERROR')
-
 def acquire_lock():
     lock = FileLock("broadcast-faircoins")
     logger.debug("acquiring lock...")
@@ -85,7 +67,7 @@ def create_address_for_agent(agent):
             entity = agent.agent_type.name,
             )
         return address
-    if init_electrum_fair():
+    if efn.is_connected():
         try:
             address = efn.new_fair_address(
                 entity_id = agent.nick.encode('ascii','ignore'),
@@ -132,7 +114,7 @@ def create_requested_addresses():
         return "failed to get FairCoin address requests"
 
     if requests:
-        if init_electrum_fair():
+        if efn.is_connected():
             logger.debug("broadcast_tx ready to process FairCoin address requests")
             for resource in requests:
                 result = create_address_for_resource(resource)
@@ -162,13 +144,9 @@ def broadcast_tx():
     try:
         successful_events = 0
         failed_events = 0
-        if events and init_electrum_fair():
+        if events and efn.is_connected():
             logger.debug("broadcast_tx ready to process events")
             for event in events:
-                #do we need to check for missing digital_currency_address here?
-                #and create them?
-                #fee = efn.network_fee() # In Satoshis
-                #fee = Decimal("%s" %fee) / FAIRCOIN_DIVISOR
                 if event.resource:
                     if event.event_type.name=="Give":
                         address_origin = event.resource.digital_currency_address
